Renamed one of the exception modules.
